test-in-docker: fix in zsh > 5.3
The way I was filtering out entries in the frameworks array stopped
working in newer versions of ZSH; it would convert the array into a
string (you could see it with `typeset -p frameworks`)

So I rewrote it.

I don't see anything in the release notes for ZSH that would explain
this and I didn't find any option that would restore this behavior.

Related: #882